requirements.txt
discord.py
llama_index
